chore(coin-evm): outsource two helpers to llc
