+++ .changeset/moody-baboons-develop.md
@@ -1,0 +1,7 @@
+---
+"@ledgerhq/coin-evm": patch
+"live-mobile": patch
+"@ledgerhq/live-common": patch
+---
+
+chore(coin-evm): move `getDefaultFeeUnit` and `getMessageProperties` to llc
