+++ libs/coin-modules/coin-evm/package.json
@@ -65,7 +65,6 @@
   "dependencies": {
     "@ethersproject/shims": "^5.7.0",
     "@ledgerhq/coin-module-framework": "catalog:",
-    "@ledgerhq/evm-tools": "workspace:^",
     "@ledgerhq/ledger-wallet-framework": "workspace:^",
     "@ledgerhq/live-env": "workspace:^",
     "@ledgerhq/live-network": "workspace:^",
