+++ libs/coin-modules/coin-evm/src/logic.test.ts
@@ -1,5 +1,4 @@
 import { getCryptoCurrencyById } from "@ledgerhq/ledger-wallet-framework/currencies";
-import * as EVM_TOOLS from "@ledgerhq/evm-tools/message/EIP712/index";
 
 jest.mock("./network/node/index", () => ({
   ...jest.requireActual("./network/node/index"),
@@ -9,7 +8,7 @@ jest.mock("./network/node/index", () => ({
 }));
 
 import { getCoinConfig } from "./config";
-import { getAdditionalLayer2Fees, getMessageProperties } from "./logic";
+import { getAdditionalLayer2Fees } from "./logic";
 import { getNodeApi } from "./network/node/index";
 import { padHexString, safeEncodeEIP55 } from "./utils";
 
@@ -159,30 +158,6 @@ describe("EVM Family", () => {
         const address = "0x00000";
         const encodedAddress = safeEncodeEIP55(address);
         expect(encodedAddress).toBe(address);
-      });
-    });
-
-    describe("getMessageProperties", () => {
-      it("should return null if the message isn't an EIP712", async () => {
-        expect(await getMessageProperties({ standard: "EIP191", message: "doot-doot" })).toBe(null);
-      });
-
-      it("should return the fields displayed on the nano", async () => {
-        jest.spyOn(EVM_TOOLS, "getEIP712FieldsDisplayedOnNano").mockResolvedValueOnce([
-          {
-            label: "key",
-            value: "value",
-          },
-        ]);
-
-        expect(
-          await getMessageProperties({
-            standard: "EIP712",
-            message: {} as any,
-            domainHash: "0xabc",
-            hashStruct: "0xdef",
-          }),
-        ).toEqual([{ label: "key", value: "value" }]);
       });
     });
   });
